+++ Model/Config/Source/ChargedCurrency.php
@@ -1,0 +1,39 @@
+<?php
+/**
+ * Altapay Module for Magento 2.x.
+ *
+ * Copyright © 2018 Altapay. All rights reserved.
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SDM\Altapay\Model\Config\Source;
+
+use Magento\Framework\Option\ArrayInterface;
+
+class ChargedCurrency implements ArrayInterface
+{
+
+    /**
+     * @var array
+     */
+    private static $chargedCurrencyOptions = [
+        'display_currency' => 'Display currency',
+        'base_currency' => 'Global/Website currency'
+    ];
+
+    /**
+     * Return array of options as value-label pairs
+     *
+     * @return array Format: array(array('value' => '<value>', 'label' => '<label>'), ...)
+     */
+    public function toOptionArray()
+    {
+        $output = [];
+        foreach (self::$chargedCurrencyOptions as $key => $label) {
+            $output[] = ['value' => $key, 'label' => $label];
+        }
+
+        return $output;
+    }
+}